requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# LangChain & AI
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.0
langchain-text-splitters==0.3.0

# Vector Store
chromadb==0.5.0

# Embeddings
sentence-transformers==3.0.0

# Memory
redis==5.0.0

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
unstructured==0.15.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0

# Testing
pytest==8.3.0
httpx==0.27.0
pytest-asyncio==0.24.0
